requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.12
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
